Fixes #8: POP_EPOCH bins problem w/ STR2DOUBLE
POP_EPOCHBIN uses commas to separate bins, however the STR2DOUBLE
function call on line 280, strips the commas out.

This wasn’t a problem before because the BIOSIG toolbox overwrote
Matlab’s builtin STR2DOUBLE. This is an issue now since the BIOSIG
toolbox doesn’t come preloaded with EEGLAB.

The fix is to uses the STR2NUM function which doesn’t not strip out
commas, and thus preserves the separation of bins.