Fixes issue #14: Synchronizing artifacts between RT and ERPs
This fixes the issue of ensuring the trials/epochs that go into the
ERPs are the same as those that are outputted as behavioral RTs.

The problem stems from the way ERPLAB marks and detects artifacts for
rejection. When ERPLAB detects an artifact for rejection, basically it
looks for the response-time event code in an epoch and uses that to
mark the response-time for rejection. In the special case where an
epoch does not encompass the response-code (i.e. when the epoch window
is too short to include the response-code), there is no response event
to mark and thus the trial is erroneously included in the behavioral RT
output. For example, in a single trial with an 1000-ms response and an
epoch window of [-200 800] the resulting epoch would only include the
time-locked stimulus event code. And so later, in artifact detection
processing ERPLAB would fail to mark the response time.