Fixed plot AutoYlim default bug
Fixes #48

When plotting ERPs, there was an occasional issue where setting the
YScale via the GUI options could give a plot with 'AutoYLim' being off,
but the equivalent script command would leave 'AutoYLim' on.

A workaround for the this was to include 'AutoYLim', 'off' in the
scripting command. We have now changed the default behaviour of
pop_ploterps() to act more as expected, and so 'AutoYLim' is taken to
be 'off' by default when Yscale is specified via script.

As such, 'AutoYLim', 'off' no longer needs to be set in scripts in
cases like this, but scripts including these arguments will still work
as expected.

This is implemented by changing autoYlim default from ‘on’ to
‘on-by-default’, and if this is on, turning this off when specific
yscale is set.