Datatype check fix
For checks of ERP.datatype, I have rearranged the string checks so that
ERP type is both the default and the fallback.

The power options should still work, as FFT, EFFT and TFFT should be
caught as power-like, and plotted appropriately.

This means we can plot new datatypes without specifying each one. The
new datatypes will be plotted with ERP-like style (unless ERP.datatype
ends with 'FFT', in which case it will be plotted like power data).