Avoid empty erpset list for MW
There was a bug where the Measurement Window GUI would occasionally not
open, complaining that numel(erpsets) == 0.

This seems to be due to a flaw in how the erpset list was sorted.

I now check for this being empty, and re-set the erpset list to just
the current erpset in that case.