EEG amplitude FFT fix
Wendy Zheng reported a bug where some portion of the EEG might not be counted toward a plotted Amplitude Spectrum.

This seems to be due to the window used for FFT not moving through the entire continuous EEG data.

This commit introduces a change to how the number of windows (nwindows) is calculated, and so should advance though to cover most of the provided EEG data.