Frac peak lat out of range
Addresses #92

This stops the fractional peak latancy from being outside the desired time range.

I rewrote the loop to find the correct latency of the closest matching point to the local peak.

It now also confirms that this point is not in the 'padding' of the neighbouring points, and returns a NaN is this is so.