Improvements to FFT, boundary check
This commit includes various improvements to the EEG FFT Amplitude tool.

There is a check for boundary events within the sample window, on by default.

There is a new function, 'find_boundary_times', to simplify this.

There is also an optional argument for log plots, off by default.

The window size can be modified, default 5 seconds.

A report is output to the command window.