Change menu postion request, fix bug
There was a little display bug in how Matlab R2020a was showing the ERPLAB and ERPset menus in the EEGLAB GUI.

This change removes the request for a specific position for the ERPLAB menus, which means these menus are consistently placed at the end. That's an improvement over being placed at the start, and messing up the menu ordering.